requirements.txt
telebot~=0.0.5
aerich~=0.7.2
click~=8.1.7
tortoise-orm
python-dotenv~=1.0.0
aiohttp